Misc refactor to NETCore 1.0.0
