+++ .gitignore
@@ -235,3 +235,5 @@ _Pvt_Extensions
 
 # FAKE - F# Make
 .fake/
+*.db
+/build/1
