+++ Flurl.Http.Xml.sln
@@ -3,21 +3,16 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
 VisualStudioVersion = 14.0.25420.1
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "Flurl.Http.Xml.Shared", "Flurl.Http.Xml.Shared\Flurl.Http.Xml.Shared.shproj", "{16A3075D-B39A-44E4-9A4F-5855392B40C7}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{F2114366-0322-434F-91D3-A25939F0CFA7}"
 	ProjectSection(SolutionItems) = preProject
 		global.json = global.json
 	EndProjectSection
 EndProject
-Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Flurl.Http.Xml.Tests", "..\test\Flurl.Http.Xml.Tests\Flurl.Http.Xml.Tests.xproj", "{7BA9BC27-1501-47EA-90FF-BCDBD25CA8B8}"
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Flurl.Http.Xml.Tests", "test\Flurl.Http.Xml.Tests\Flurl.Http.Xml.Tests.xproj", "{7BA9BC27-1501-47EA-90FF-BCDBD25CA8B8}"
 EndProject
-Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Flurl.Http.Xml", "Flurl.Http.Xml\Flurl.Http.Xml.xproj", "{B9826CF6-17DD-4D2D-A6BD-F15CFA18F9CB}"
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Flurl.Http.Xml", "src\Flurl.Http.Xml\Flurl.Http.Xml.xproj", "{B9826CF6-17DD-4D2D-A6BD-F15CFA18F9CB}"
 EndProject
 Global
-	GlobalSection(SharedMSBuildProjectFiles) = preSolution
-		Flurl.Http.Xml.Shared\Flurl.Http.Xml.Shared.projitems*{16a3075d-b39a-44e4-9a4f-5855392b40c7}*SharedItemsImports = 13
-	EndGlobalSection
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Release|Any CPU = Release|Any CPU
