+++ appveyor.yml
@@ -1,27 +1,17 @@
-version: 1.2.0.{build}
+version: 1.3.0.{build}
 branches:
   only:
   - master
 pull_requests:
   do_not_increment_build_number: true
-configuration: Release
-platform: Any CPU
-assembly_info:
-  patch: true
-  file: '**\AssemblyInfo.*'
-  assembly_version: '{version}'
-  assembly_file_version: '{version}'
-  assembly_informational_version: '{version}'
-before_build:
-- cmd: >-
-    cd src
-
-    nuget restore
-
-    cd ..
-build:
-  publish_nuget: false
-  verbosity: minimal
+install:
+  - ps: .\build\install.ps1
+build_script:
+- ps: .\build\build.ps1 $env:APPVEYOR_BUILD_VERSION $env:APPVEYOR_REPO_TAG_NAME
+test_script:
+- ps: .\build\test.ps1
+artifacts:
+- path: '**\*.nupkg'
 deploy:
 - provider: NuGet
   api_key:
