--- src/.vs/restore.dg
@@ -1,1 +1,0 @@
-#:C:\gitrepo\Flurl.Http.Xml\feature.vnext\src\Flurl.Http.Xml\Flurl.Http.Xml.xproj
