--- src/Flurl.Http.Xml.Shared/Flurl.Http.Xml.Shared.projitems
@@ -1,22 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <MSBuildAllProjects>$(MSBuildAllProjects);$(MSBuildThisFileFullPath)</MSBuildAllProjects>
-    <HasSharedItems>true</HasSharedItems>
-    <SharedGUID>16a3075d-b39a-44e4-9a4f-5855392b40c7</SharedGUID>
-  </PropertyGroup>
-  <PropertyGroup Label="Configuration">
-    <Import_RootNamespace>Flurl.Http.Xml.Shared</Import_RootNamespace>
-  </PropertyGroup>
-  <ItemGroup>
-    <Compile Include="$(MSBuildThisFileDirectory)CapturedXmlContent.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)FlurlClientExtensions.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)FlurlHttpSettingsExtensions.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)HttpResponseMessageExtensions.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)MicrosoftXmlSerializer.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)StringExtensions.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)UrlExtensions.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Utf8StringWriter.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)XDocumentExtensions.cs" />
-  </ItemGroup>
-</Project>
