--- src/Flurl.Http.Xml.nuspec
@@ -1,52 +1,0 @@
-<?xml version="1.0"?>
-<package >
-  <metadata>
-    <id>Flurl.Http.Xml</id>
-    <version>$version$</version>
-    <title>XML extensions to Flurl.Http</title>
-    <authors>Luk Vermeulen</authors>
-    <owners>Luk Vermeulen</owners>
-    <requireLicenseAcceptance>true</requireLicenseAcceptance>
-    <description>XML extensions to Flurl.Http</description>
-    <copyright>Copyright 2016</copyright>
-    <projectUrl>https://github.com/lvermeulen/Flurl.Http.Xml</projectUrl>
-    <licenseUrl>https://github.com/lvermeulen/Flurl.Http.Xml/blob/master/LICENSE</licenseUrl>
-    <iconUrl>http://i.imgur.com/llEKpRL.png?1</iconUrl>
-    <tags>Flurl Http Xml</tags>
-    <dependencies>
-      <group targetFramework="net451">
-        <dependency id="Flurl.Http" version="1.0.0-beta7" />
-      </group>
-      <group targetFramework="netstandard1.4">
-        <!--<dependency id="NETStandard.Library" version="1.6.0" />-->
-        <dependency id="System.Dynamic.Runtime" version="4.0.11" />
-        <dependency id="System.Net.Http" version="4.1.0" />
-        <dependency id="System.Text.RegularExpressions" version="4.1.0" />
-        <dependency id="System.Threading" version="4.0.11" />
-        <dependency id="System.Xml.XDocument" version="4.0.11" />
-        <dependency id="System.Xml.XmlSerializer" version="4.0.11" />
-        <dependency id="System.Xml.XPath" version="4.0.1" />
-        <dependency id="System.Xml.XPath.XDocument" version="4.0.1" />
-        <dependency id="Flurl.Http" version="1.0.0-beta7" />
-      </group>
-      <group targetFramework="netstandard1.6">
-        <!--<dependency id="NETStandard.Library" version="1.6.0" />-->
-        <dependency id="System.Dynamic.Runtime" version="4.0.11" />
-        <dependency id="System.Net.Http" version="4.1.0" />
-        <dependency id="System.Text.RegularExpressions" version="4.1.0" />
-        <dependency id="System.Threading" version="4.0.11" />
-        <dependency id="System.Xml.XDocument" version="4.0.11" />
-        <dependency id="System.Xml.XmlSerializer" version="4.0.11" />
-        <dependency id="System.Xml.XPath" version="4.0.1" />
-        <dependency id="System.Xml.XPath.XDocument" version="4.0.1" />
-        <dependency id="Flurl.Http" version="1.0.0-beta7" />
-        <dependency id="Flurl.Http" version="1.0.0-beta7" />
-      </group>
-    </dependencies>  
-  </metadata>
-  <files>
-    <file src="..\Flurl.Http.Xml\bin\Release\net451\Flurl.Http.Xml.dll" target="lib\net451" />
-    <file src="..\Flurl.Http.Xml\bin\Release\netstandard1.4\Flurl.Http.Xml.dll" target="lib\netstandard1.4" />
-    <file src="..\Flurl.Http.Xml\bin\Release\netstandard1.6\Flurl.Http.Xml.dll" target="lib\netstandard1.6" />
-  </files>
-</package>
