+++ src/Flurl.Http.Xml/project.json
@@ -1,17 +1,13 @@
 ﻿{
   "title": "Flurl.Http.Xml",
-  "version": "1.0.0-*",
+  "version": "1.0.0",
   "authors": [ "Luk Vermeulen" ],
   "copyright": "Copyright 2016",
   "description": "XML extensions to Flurl.Http",
 
   "buildOptions": {
-    "compile": {
-      "include": "../Flurl.Http.Xml.Shared/**/*.cs"
-    },
     "optimize": true,
-    "outputName": "Flurl.Http.Xml",
-    "xmlDoc": false
+    "xmlDoc": true
   },
 
   "packOptions": {
@@ -33,43 +29,16 @@
   },
 
   "frameworks": {
-    "net451": {
+    "net45": {
       "frameworkAssemblies": {
-        "System.Dynamic.Runtime": "4.0.0",
-        "System.Linq": "4.0.0",
         "System.Net.Http": "4.0.0",
-        "System.Text.RegularExpressions": "4.0.0",
-        "System.Threading": "4.0.0",
         "System.Xml": "4.0.0.0",
         "System.Xml.Linq": "4.0.0.0"
-      },
-      "dependencies": {
-        "NETStandard.Library": "1.6.0"
       }
     },
     "netstandard1.4": {
       "dependencies": {
-        "NETStandard.Library": "1.6.0",
-        "System.Dynamic.Runtime": "4.0.11",
-        "System.Net.Http": "4.1.0",
-        "System.Text.RegularExpressions": "4.1.0",
-        "System.Threading": "4.0.11",
-        "System.Xml.XDocument": "4.0.11",
         "System.Xml.XmlSerializer": "4.0.11",
-        "System.Xml.XPath": "4.0.1",
-        "System.Xml.XPath.XDocument": "4.0.1"
-      }
-    },
-    "netstandard1.6": {
-      "dependencies": {
-        "NETStandard.Library": "1.6.0",
-        "System.Dynamic.Runtime": "4.0.11",
-        "System.Net.Http": "4.1.0",
-        "System.Text.RegularExpressions": "4.1.0",
-        "System.Threading": "4.0.11",
-        "System.Xml.XDocument": "4.0.11",
-        "System.Xml.XmlSerializer": "4.0.11",
-        "System.Xml.XPath": "4.0.1",
         "System.Xml.XPath.XDocument": "4.0.1"
       }
     }
