+++ test/Flurl.Http.Xml.Tests/MicrosoftXmlSerializerShould.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
+﻿using System.Text;
 using System.Xml;
 using Flurl.Http.Xml.Tests.Models;
 using Xunit;
